+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 ## 2.11.7 2024-09-12
 
 * Update dxda to 0.6.2 and dxfuse to 1.4.0
+* Added [DIY Docker image instructions](./scripts/docker_image/). This replaces the Docker image published to Docker Hub, which is deprecated and will be removed after November 1, 2024.
 
 ## 2.11.6 2024-03-26